Updated linter configuration
This commit updates the linter configuration for this project to be managed
locally. This has several benefits, but the most important for the purpose of
this commit is that prettier and eslint agree and one will not undo changes
from the other.

The babel parser is used to inform eslint of advance features and to help
eslint understand some typescript syntax. There are two main parsers for eslint
and typescript, typescript-eslint and babel, but unfortunately, they are
"currently not intended to be used together" according to the typescript-eslint
README. The practical consequences of this are that the babel parser does not
understand all type information and eslint will have a few false positives
which must be ignored on an individual basis. It is also true that babel does
not perform typechecking, but the typescript compiler itself (through tsserver)
performs this checking and when configured properly with an editor will inform
a developer about type mismatches separately.