Fixed NarrativeList Load more button
This commit fixes the button for loading more items in the narrative list in
situations where there is a prefix for the dashboard app. This is a common
feature of KBase environments and can be dealt with in most cases by keeping
links relative. Given that this approach means that the client URL will always
reflect how many narratives the user wishes to see, the list can be derived
entirely from the URL any time it updates. This eliminates the need for some
component state management callbacks which have been removed. The default user
picture is an example where absolute links are needed, so this update fixes
these references in other environments as well.