Fixed SCT-2780: Added new sorts and client cache.
This commit adds lexicographical sorting to the possible NarrativeList sorts as
well as caching to avoid repeating the same HTTP requests. Narrative previews
are also cached. This caching is only done in memory and will be lost if the
client navigates away from the page. Even so, this caching is designed so that
future versions may use string based key-value stores such as localStorage.