Fixed SCT-2781 and SCT-2782 functionality.
This commit fixes SCT-2781 by replacing usernames with real names, and in the
narrative detail these names are links to the user's profile. These links have
the username as their title attribute so that it may be seen on hover. These
usernames are also cached to prevent extraneous HTTP requests. It also fixes
SCT-2782 by identifying the most common data types within a narrative. Further,
for shared narratives only ten users are to show up by default, and the markup
for this is included in this commit, but the functionality is mediated by CSS
in a subsequent commit.