Altered search behavior when changing tabs.
This commit Fixes SCT-2877 by clearing the search term when a user changes
tabs. There are other ways to ensure the consistency of the user's experience,
such as maintaining the term and continuing to apply the search, but this is
the smallest change that hopefully will eliminate user confusion and reduce
user surprise. It may be a surprise to remove the search term, but it is
confusing to have a term present and not applied.